move the comments into a bottom sheet
