+++ app/src/main/res/layout/fragment_player.xml
@@ -347,15 +347,6 @@
                 android:descendantFocusability="blocksDescendants">
 
                 <androidx.recyclerview.widget.RecyclerView
-                    android:id="@+id/comments_recView"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="10dp"
-                    android:layout_marginBottom="10dp"
-                    android:nestedScrollingEnabled="false"
-                    android:visibility="gone" />
-
-                <androidx.recyclerview.widget.RecyclerView
                     android:id="@+id/related_rec_view"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
